LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/position-change-heuristic-ib-split-expected.txt

Some title
Sticky header
Some actual content.

FAIL Scroll offset adjustments are correctly suppressed when changing the position of an inline assert_equals: Element should become and remain fixed expected "fixed" but got "static"
